pyproject.toml
[project]
name = "deepseek-ptu-perf-test"
version = "0.1.0"
description = "Azure AI DeepSeek Performance Benchmark Test"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
]
